Fixes problem with tracebacks.
The relative path to the module will be stored as the module's pathh.
This means that if the current directory has changed by the time the
traceback is printed, it tries to read from a file that's not there and
ends up not displaying the lines of code for the user's module.

This fixes #12.