Added handling for binary data in pipelines.
I don't think this fully solves the problem and I'd like to figure out
a better way to do this. I think a way to pass in a function that
describes how to open the output file would be best (maybe the input
file too).

Test Plan:
`tox`, as well as testing with my blog.